Consistently use std::pair<double, double> limits
There are a number of places in tach sensor where we jump from having a
pair of doubles, so having a pair of ints, then back.  All sensors take
this as a double, so we should be consistent, even if tachsensor is only
expected to have integer units.

This was flagged by clang-tidy as part of the
bugprone-misplaced-widening-cast check.

+++ src/TachSensor.cpp
@@ -46,7 +46,7 @@ TachSensor::TachSensor(const std::string& path, const std::string& objectType,
                        boost::asio::io_service& io, const std::string& fanName,
                        std::vector<thresholds::Threshold>&& thresholdsIn,
                        const std::string& sensorConfiguration,
-                       const std::pair<size_t, size_t>& limits,
+                       const std::pair<double, double>& limits,
                        const PowerState& powerState,
                        const std::optional<std::string>& ledIn) :
     Sensor(escapeName(fanName), std::move(thresholdsIn), sensorConfiguration,
